==> notebook/04_housing_data_enriched_icpsr.ipynb <==
Data link: 
https://www.openicpsr.org/openicpsr/project/208462/version/V1/view?path=/openicpsr/208462/fcr:versions/V1/BJK_County_Tax_Panel_public.dta&type=file

import pandas as pd

### 2018-2022 Tax Info Sales + Income

df_tax_raw = pd.read_stata('../data/raw/BJK_State_Tax_Panel_public.dta')
df_tax_raw.head()

df_tax_cleaned = (
    df_tax_raw.loc[df_tax_raw['year'].between(2018, 2023), [
        'state', 'stateabbr', 'year', 'sales_rate_state', 'top_inc_rate_state'
    ]]
    .rename(columns={
        'state': 'StateFull',
        'stateabbr': 'StateName',
        'year': 'Year',
        'sales_rate_state': 'Sales_Tax',
        'top_inc_rate_state': 'Income_Tax'
    }).reset_index(drop=True)
)

# print(df_tax_cleaned.info())
df_tax_cleaned['StateName'].unique()

df_main = pd.read_csv('../data/processed/03_housing_data_enriched_fred_fema.csv')
df_main.head()

df_main = df_main.merge(df_tax_cleaned, on=['StateName', 'Year'], how='left')
df_main.head()

### 2018-2022 Tax Info Property

# df_tax_raw = pd.read_stata('../data/raw/BJK_County_Tax_Panel_public.dta')
# df_tax_raw.head()

# df_tax_cleaned = (
#     df_tax_raw.loc[df_tax_raw['year'].between(2018, 2023), [
#         'state', 'year', 'prop_rate'
#     ]]
#     .rename(columns={
#         'state': 'StateName',
#         'year': 'Year',
#         'prop_rate': 'Property_Tax'
#     }).reset_index(drop=True)
# )

# # print(df_tax_cleaned.info())
# df_tax_cleaned['StateName'].unique()

# df_main = df_main.merge(df_tax_cleaned, on=['StateName', 'Year'], how='left')
# df_main.head()

### 2023 Tax Info

df_raw_prop = pd.read_csv('../data/raw/Property Taxes by State and County, 2025  Tax Foundation Maps.csv')
df_raw_prop = df_raw_prop.drop(['Effective Tax Rate (2022)', 'Rank'], axis=1)
columns = {'State': 'StateFull', 'Effective Tax Rate (2023)': 'Property_Tax'}

df_raw_prop = df_raw_prop.rename(columns=columns)
df_raw_prop['Year'] = '2023'

df_prop_2023_cleaned = df_raw_prop.copy()
df_prop_2023_cleaned.head()

df_raw_sales = pd.read_excel("../data/raw/LOST_Jan_2023_table.xlsx", skiprows=3)

df_raw_sales.columns = [
    'State_Abbr', 'State_Tax_Rate', 'State_Rank',
    'Avg_Local_Tax_Rate', 'Combined_Rate',
    'Combined_Rank', 'Max_Local_Tax_Rate'
]

df_raw_sales = df_raw_sales.dropna(subset=['State_Tax_Rate'])

abbr_to_full = {
    'Ala.': 'Alabama', 'Alaska': 'Alaska', 'Ariz.': 'Arizona', 'Ark.': 'Arkansas', 'Calif.': 'California',
    'Colo.': 'Colorado', 'Conn.': 'Connecticut', 'Del.': 'Delaware', 'Fla.': 'Florida', 'Ga.': 'Georgia',
    'Hawaii': 'Hawaii', 'Idaho': 'Idaho', 'Ill.': 'Illinois', 'Ind.': 'Indiana', 'Iowa': 'Iowa',
    'Kans.': 'Kansas', 'Ky.': 'Kentucky', 'La.': 'Louisiana', 'Maine': 'Maine', 'Md.': 'Maryland',
    'Mass.': 'Massachusetts', 'Mich.': 'Michigan', 'Minn.': 'Minnesota', 'Miss.': 'Mississippi', 'Mo.': 'Missouri',
    'Mont.': 'Montana', 'Nebr.': 'Nebraska', 'Nev.': 'Nevada', 'N.H.': 'New Hampshire', 'N.J.': 'New Jersey',
    'N.M.': 'New Mexico', 'N.Y.': 'New York', 'N.C.': 'North Carolina', 'N.D.': 'North Dakota', 'Ohio': 'Ohio',
    'Okla.': 'Oklahoma', 'Ore.': 'Oregon', 'Pa.': 'Pennsylvania', 'R.I.': 'Rhode Island', 'S.C.': 'South Carolina',
    'S.D.': 'South Dakota', 'Tenn.': 'Tennessee', 'Tex.': 'Texas', 'Utah': 'Utah', 'Vt.': 'Vermont',
    'Va.': 'Virginia', 'Wash.': 'Washington', 'W.Va.': 'West Virginia', 'Wis.': 'Wisconsin', 'Wyo.': 'Wyoming',
    'D.C.': 'District of Columbia'
}

df_raw_sales['StateFull'] = df_raw_sales['State_Abbr'].map(abbr_to_full)

df_sales_2023_cleaned = df_raw_sales[['StateFull', 'Combined_Rate']].copy()
df_sales_2023_cleaned['Year'] = '2023'
df_sales_2023_cleaned = df_sales_2023_cleaned.rename(columns={'Combined_Rate': 'Sales_Tax'})

df_sales_2023_cleaned.head()

# Save merged DataFrame
df_main.to_csv('../data/processed/04_housing_data_enriched_icpsr.csv', index=False)

print("File successfully saved to 'data/processed/04_housing_data_enriched_icpsr.csv'")
