==> notebook/05_housing_data_renamed_columns.ipynb <==
import pandas as pd
import re

df = pd.read_csv('../data/processed/04_housing_data_enriched_icpsr.csv')
df.head()

df.columns

rename_dict = {
    "RegionID": "region_id",
    "RegionName": "region_name",
    "RegionType": "region_type",
    "StateName": "state_name",
    "SizeRank": "size_rank",
    "Date": "date",
    "ZHVI_AllHomes": "zhvi_all_homes",
    "ZHVI_SingleFamily": "zhvi_single_family",
    "ZHVI_Condo": "zhvi_condo",
    "SalesCount_SFR": "sales_count_sfr",
    "Inventory_SFR": "inventory_sfr",
    "Year": "year",
    "Month": "month",
    "Unemployment_Rate": "unemployment_rate",
    "Median_Income": "median_income",
    "Population": "population",
    "Num_Major_Disasters_Last_12mo": "num_major_disasters_last_12mo",
    "Major_Disaster_Declared": "major_disaster_declared",
    "Most_Recent_Disaster_Type": "most_recent_disaster_type",
    "StateFull": "state_full",
    "Sales_Tax": "sales_tax",
    "Income_Tax": "income_tax"
}

df.rename(columns=rename_dict, inplace=True)

df.columns

# Save renamed DataFrame
df.to_csv('../data/processed/05_housing_data_renamed_columns.csv', index=False)

print("File successfully saved to 'data/processed/05_housing_data_renamed_columns.csv'")
